// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  STUDENT: 'STUDENT',
  ADMIN: 'ADMIN',
  TEACHER: 'TEACHER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const QuestionType = {
  TRUE_FALSE: 'TRUE_FALSE',
  MULTIPLE_CHOICE: 'MULTIPLE_CHOICE',
  MULTI_SELECT: 'MULTI_SELECT',
  COMPLETE: 'COMPLETE',
  WRITTEN: 'WRITTEN'
} as const

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType]


export const LessonType = {
  VIDEO: 'VIDEO',
  TEXT: 'TEXT',
  MATERILAS: 'MATERILAS'
} as const

export type LessonType = (typeof LessonType)[keyof typeof LessonType]


export const ChatRole = {
  USER: 'USER',
  MODEL: 'MODEL'
} as const

export type ChatRole = (typeof ChatRole)[keyof typeof ChatRole]